refactor : 통합 테스트 DisplayName 수정

+++ src/main/java/softwar7/application/schedule/ScheduleDeleteService.java
@@ -3,7 +3,11 @@ package softwar7.application.schedule;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import softwar7.domain.schedule.persist.Schedule;
+import softwar7.global.constant.ExceptionMessage;
+import softwar7.global.exception.ForbiddenException;
 import softwar7.repository.schedule.ScheduleRepository;
+
+import static softwar7.global.constant.ExceptionMessage.*;
 
 @Service
 public class ScheduleDeleteService {
@@ -14,8 +18,11 @@ public class ScheduleDeleteService {
     }
 
     @Transactional
-    public void deleteSchedule(final long scheduleId){
+    public void deleteSchedule(final long memberId, final long scheduleId){
         Schedule schedule = scheduleRepository.getById(scheduleId);
+        if (memberId != schedule.getMemberId()) {
+            throw new ForbiddenException(SCHEDULE_FORBIDDEN_EXCEPTION.message);
+        }
         scheduleRepository.delete(schedule);
     }
 }
